pyproject.toml
[project]
name = "women_in_mathematics"
requires-python = ">=3.11"  
version = "0.1.0"
dependencies = [
    "dagster==1.12.2",
    "dateparser>=1.2.2",
    "json-repair>=0.53.0",
    "matplotlib>=3.10.7",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pdfminer.six>=20231228",
    "pymupdf>=1.26.6",
    "pypdf2>=3.0.1",
    "pyprojroot>=0.3.0",
    "seaborn>=0.13.2",
    "storywrangler-sdk",
    "zodiac-sign>=0.2.5",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "women_in_mathematics"
registry_modules = [
    "women_in_mathematics.components.*",
]

[tool.uv.sources]
storywrangler-sdk = { path = "../storywrangler/packages/sdk" }


